refactor: address second-pass review findings
- Narrow broad 'except Exception' in rosdep_validator, verify_rosdep_mapping,
  and workspace; delete dead bare except after the narrow catch in
  'can_resolve'. Top-level CLI guard in verify_rosdep_mapping keeps
  'except Exception' (documented; tests assert permissive behaviour).
- Expose 'rospkg.ResourceNotFound' via rosdep_wrapper; whitelist 'rospkg.*'
  in mypy.
- Delete dead 'check_count' wrap in package_xml_validator.
- Split the 750-line validation_steps.py into a package, one step per file.
  Public names re-exported from __init__.py — call sites unchanged.
- Enable branch coverage in tests/coverage_test.py. New baseline: 89% line /
  89% branch (was 92% / 0% — branch wasn't tracked).
- Register 'validation_steps' and 'formatter' subpackages in pyproject.
