+++ package_xml_validation/helpers/validation_steps/rosdep_check_step.py
@@ -1,0 +1,51 @@
+"""Rosdep validation step."""
+
+from __future__ import annotations
+
+import os
+from typing import TYPE_CHECKING
+
+from ._base import ValidationConfig, ValidationResult, ValidationStep
+
+if TYPE_CHECKING:
+    from ..package_types import XmlElement
+    from ..pkg_xml_formatter import PackageXmlFormatter
+    from ..rosdep_validator import RosdepValidator
+
+
+class RosdepCheckStep(ValidationStep):
+    name = "ROS dependency check"
+
+    def __init__(
+        self,
+        config: ValidationConfig,
+        formatter: PackageXmlFormatter,
+        rosdep_validator: RosdepValidator,
+    ) -> None:
+        """Initialize ROS dependency validation step."""
+        super().__init__(config)
+        self.formatter = formatter
+        self.rosdep_validator = rosdep_validator
+
+    def perform_check(self, root: XmlElement, xml_file: str) -> ValidationResult:
+        """Validate rosdep keys in package.xml dependencies."""
+        result = ValidationResult(root=root)
+        if not self.config.check_rosdeps or self.config.missing_deps_only:
+            return result
+
+        rosdeps = self.formatter.retrieve_all_dependencies(root)
+        if not rosdeps:
+            return result
+
+        unresolvable = self.rosdep_validator.check_rosdeps_and_local_pkgs(rosdeps)
+        if not unresolvable:
+            return result
+
+        pkg_name = os.path.basename(os.path.dirname(xml_file))
+        message = (
+            f"Unresolvable ROS dependencies found in {pkg_name}/package.xml: "
+            f"{', '.join(unresolvable)}"
+        )
+        result.critical_errors.append(message)
+        result.valid = False
+        return result
